meetings-backend/pyproject.toml
[tool.poetry]
name = "meetings-backend"
version = "0.1.0"
description = ""
authors = ["Diego Alvarez S. <dialvarezs@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <3.11"
fastapi = "^0.70.1"
uvicorn = { extras = ["standard"], version = "^0.16.0" }
sqlmodel = "^0.0.6"
alembic = "^1.7.5"
psycopg2-binary = "^2.9.3"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["argon2"], version = "^1.7.4" }
python-multipart = "^0.0.5"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]
poethepoet = "^0.11.0"
black = "^21.12b0"
isort = "^5.10.1"
pytest = "^6.2.5"
requests = "^2.26.0"

[tool.poe]
envfile = ".env"

[tool.poe.tasks]
_isort = "isort app/"
_black = "black app/"
shell = "poetry shell"
start = { shell = "uvicorn app.main:app --reload" }
test = { shell = "pytest", env = { "PYTHONPATH" = "." } }
format = ["_isort", "_black"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
